persist authCode
