+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/coronatest/type/pcr/PCRCoronaTest.kt
@@ -19,6 +19,9 @@ data class PCRCoronaTest(
     @SerializedName("registrationToken")
     override val registrationToken: RegistrationToken,
 
+    @SerializedName("authCode")
+    override val authCode: String? = null,
+
     @SerializedName("isSubmitted")
     override val isSubmitted: Boolean = false,
