+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/coronatest/type/rapidantigen/RACoronaTest.kt
@@ -28,6 +28,9 @@ data class RACoronaTest(
     @SerializedName("registrationToken")
     override val registrationToken: RegistrationToken,
 
+    @SerializedName("authCode")
+    override val authCode: String? = null,
+
     @SerializedName("isSubmitted")
     override val isSubmitted: Boolean = false,
