+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/coronatest/type/rapidantigen/RATestProcessor.kt
@@ -165,6 +165,13 @@ class RATestProcessor @Inject constructor(
         }
     }
 
+    override suspend fun updateAuthCode(test: PersonalCoronaTest, authCode: String): PersonalCoronaTest {
+        Timber.tag(TAG).v("updateAuthCode(test=%s)", test)
+        test as RACoronaTest
+
+        return test.copy(authCode = authCode)
+    }
+
     override suspend fun markSubmitted(test: PersonalCoronaTest): RACoronaTest {
         Timber.tag(TAG).d("markSubmitted(test=%s)", test)
         test as RACoronaTest
