--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/playbook/PlaybookModule.kt
@@ -1,13 +1,0 @@
-package de.rki.coronawarnapp.playbook
-
-import dagger.Module
-import dagger.Provides
-import javax.inject.Singleton
-
-@Module
-class PlaybookModule {
-
-    @Singleton
-    @Provides
-    fun providePlaybook(defaultPlayBook: DefaultPlaybook): Playbook = defaultPlayBook
-}
